reports/serializers.py
from rest_framework import serializers
from .models import Report


class ReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = Report
        fields = ['id', 'report_id', 'report_type', 'status', 'start_date', 
                  'end_date', 'file_url', 'created_at', 'completed_at']
        read_only_fields = ['created_at', 'completed_at']
